% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeNPT.r
\name{randomizeNPT}
\alias{randomizeNPT}
\title{Remove any objects in the workspace
remove verbose warnings
**************************************************************************#
********************* Manually updated section ***************************#
PART I: Obtain working directory containing the Excel file 
Ensure that for entries>12 generate the randomization files using ALPHA
program and copy them in this directory}
\usage{
randomizeNPT(DIR = "", DSN = "", File = "")
}
\description{
Remove any objects in the workspace
remove verbose warnings
**************************************************************************#
********************* Manually updated section ***************************#
PART I: Obtain working directory containing the Excel file 
Ensure that for entries>12 generate the randomization files using ALPHA
program and copy them in this directory
}

